+++ crinaed/app/src/main/java/com/example/crinaed/database/entity/CourseBought.java
@@ -6,8 +6,8 @@ import java.util.Date;
 
 @Entity(primaryKeys = {"idUser", "idCourse"})
 public class CourseBought {
-    public int idUser;
-    public int idCourse;
+    public long idUser;
+    public long idCourse;
     public int level;
     public float progression;
     public Date purchaseDate;
