+++ crinaed/app/src/main/java/com/example/crinaed/database/entity/UserLevel.java
@@ -7,7 +7,7 @@ import static androidx.room.ForeignKey.CASCADE;
 
 @Entity(primaryKeys = {"idUser", "cat"})
 public class UserLevel {
-    public int idUser;
+    public long idUser;
     public String cat;
     public int PE;
     public int level;
